Refactor Solana transaction input handling and improve modularity
* Replace the existing transaction serialization logic with a new function for generating Solana transaction input data.
* Move the transaction input data logic to a dedicated module for better organization and reusability.
* Remove deprecated code related to transaction input handling, enhancing clarity and maintainability.
* Update the Solana transaction scanning resolver to utilize the new input data function, streamlining the process.
